+++ test/DynamoCoreWpfTests/DynamoCoreWpfTests.csproj
@@ -131,6 +131,7 @@
     <Compile Include="..\..\src\AssemblySharedInfoGenerator\AssemblySharedInfo.cs">
       <Link>AssemblySharedInfo.cs</Link>
     </Compile>
+    <Compile Include="NodeViewTests.cs" />
     <Compile Include="NodeExecutionUITest.cs" />
     <Compile Include="CustomNodeTests.cs" />
     <Compile Include="DateTimeTests.cs" />
