+++ test/DynamoCoreWpfTests/DynamoTestUIBase.cs
@@ -6,9 +6,11 @@ using System.Reflection;
 using System.Threading;
 using Dynamo.Configuration;
 using Dynamo.Controls;
+using Dynamo.Graph.Nodes;
 using Dynamo.Models;
 using Dynamo.Scheduler;
 using Dynamo.ViewModels;
+using DynamoCoreWpfTests.Utility;
 using DynamoShapeManager;
 using NUnit.Framework;
 using TestServices;
@@ -178,6 +180,24 @@ namespace DynamoCoreWpfTests
 
             if (!Directory.Exists(TempFolder))
                 Directory.CreateDirectory(TempFolder);
+        }
+
+        public NodeView NodeViewOf<T>() where T : NodeModel
+        {
+            var nodeViews = View.NodeViewsInFirstWorkspace();
+            var nodeViewsOfType = nodeViews.OfNodeModelType<T>();
+            Assert.AreEqual(1, nodeViewsOfType.Count(), "Expected a single NodeView of provided type in the workspace!");
+
+            return nodeViewsOfType.First();
+        }
+
+        public NodeView NodeViewWithGuid(string guid)
+        {
+            var nodeViews = View.NodeViewsInFirstWorkspace();
+            var nodeViewsOfType = nodeViews.Where(x => x.ViewModel.NodeLogic.GUID.ToString() == guid);
+            Assert.AreEqual(1, nodeViewsOfType.Count(), "Expected a single NodeView with guid: " + guid);
+
+            return nodeViewsOfType.First();
         }
 
         #endregion
